RED-156238 Implement lazy load subscribe to key space. called once we create a new timeserise (#1741)
